# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vawulence City Media API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vote-cards/{voteCardId}/comments:
    get:
      operationId: listVoteCardComments
      tags: [comments]
      summary: List comments for a vote card
      parameters:
        - name: voteCardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createVoteCardComment
      tags: [comments]
      summary: Post a comment on a vote card
      parameters:
        - name: voteCardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ── Gists ─────────────────────────────────────────────────────────────────────
...
  /vote-cards:
    get:
      operationId: listVoteCards
      tags: [vote-cards]
      summary: List active vote cards
      parameters:
        - name: all
          in: query
          schema:
            type: boolean
          description: If true, include inactive cards (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VoteCard"
    post:
      operationId: createVoteCard
      tags: [vote-cards]
      summary: Create a vote card (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteCardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteCard"
  /vote-cards/{id}:
    get:
      operationId: getVoteCard
      tags: [vote-cards]
      summary: Get a single vote card with user vote status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteCardDetail"
        "404":
          description: Not found
    patch:
      operationId: updateVoteCard
      tags: [vote-cards]
      summary: Update a vote card (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteCardUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteCard"
    delete:
      operationId: deleteVoteCard
      tags: [vote-cards]
      summary: Delete a vote card (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /vote-cards/{id}/vote:
    post:
      operationId: castVote
      tags: [vote-cards]
      summary: Cast a vote on a vote card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteCardVoteInput"
      responses:
        "200":
          description: Updated vote card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteCardDetail"
        "409":
          description: Already voted

  # ── Trending ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Comment:
      type: object
      required: [id, content, likeCount, createdAt]
      properties:
        id:              { type: integer }
        postId:          { type: ["integer", "null"] }
        voteCardId:      { type: ["integer", "null"] }
        userId:          { type: ["string", "null"] }
        userName:        { type: ["string", "null"] }
        parentCommentId: { type: ["integer", "null"] }
        content:         { type: string }
        likeCount:       { type: integer }
        createdAt:       { type: string }
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    CommentInput:
      type: object
      required: [content, userId]
      properties:
        content:         { type: string }
        userId:          { type: string }
        parentCommentId: { type: ["integer", "null"] }

    # ── Gist ──────────────────────────────────────────────────────────────────
...
    VoteCard:
      type: object
      required: [id, title, optionALabel, optionBLabel, optionACount, optionBCount, isActive, createdAt, totalVotes]
      properties:
        id:           { type: integer }
        title:        { type: string }
        imageUrl:     { type: ["string", "null"] }
        imageUrl2:    { type: ["string", "null"] }
        optionALabel: { type: string }
        optionBLabel: { type: string }
        optionACount: { type: integer }
        optionBCount: { type: integer }
        isActive:     { type: boolean }
        createdAt:    { type: string }
        totalVotes:   { type: integer }
        commentCount: { type: integer }
    VoteCardDetail:
      type: object
      required: [id, title, optionALabel, optionBLabel, optionACount, optionBCount, isActive, createdAt, totalVotes, userVote]
      properties:
        id:           { type: integer }
        title:        { type: string }
        imageUrl:     { type: ["string", "null"] }
        imageUrl2:    { type: ["string", "null"] }
        optionALabel: { type: string }
        optionBLabel: { type: string }
        optionACount: { type: integer }
        optionBCount: { type: integer }
        isActive:     { type: boolean }
        createdAt:    { type: string }
        totalVotes:   { type: integer }
        commentCount: { type: integer }
        userVote:     { type: ["string", "null"] }
    VoteCardInput:
      type: object
      required: [title, optionALabel, optionBLabel]
      properties:
        title:        { type: string }
        imageUrl:     { type: string }
        imageUrl2:    { type: string }
        optionALabel: { type: string }
        optionBLabel: { type: string }
    VoteCardUpdate:
      type: object
      properties:
        title:        { type: string }
        imageUrl:     { type: string }
        imageUrl2:    { type: string }
        optionALabel: { type: string }
        optionBLabel: { type: string }
        isActive:     { type: boolean }
    VoteCardVoteInput:
      type: object
      required: [userId, chosenOption]
      properties:
        userId:       { type: string }
        chosenOption: { type: string, enum: [a, b] }

    # ── Trending ──────────────────────────────────────────────────────────────
...