# lib/api-zod/src/generated/types/voteCard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vawulence City Media API
 * OpenAPI spec version: 0.1.0
 */

export interface VoteCard {
  id: number;
  title: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageUrl2?: string | null;
  optionALabel: string;
  optionBLabel: string;
  optionACount: number;
  optionBCount: number;
  isActive: boolean;
  createdAt: string;
  totalVotes: number;
  commentCount?: number;
}