# lib/api-zod/src/generated/types/voteCardVoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vawulence City Media API
 * OpenAPI spec version: 0.1.0
 */
import type { VoteCardVoteInputChosenOption } from './voteCardVoteInputChosenOption';

export interface VoteCardVoteInput {
  userId: string;
  chosenOption: VoteCardVoteInputChosenOption;
}